Refactor the S3 functionality into the SimpleS3BucketManager
The intention of this change is to simplify the handlers so that they
focus on what they're meant to do, which is pull relevant details from
the request, pass said details to code that handles the details of the
request, such as the S3 manager, and then provides a response, based on
the result of the relevant operations on the request data.

So, to me, this change moves the S3 functionality to the manager type,
so that the handlers don't know too much or attempt to do too much.